+++ packages/web3-providers/src/Chainbase/apis/ChainbaseHistoryAPI.ts
@@ -1,0 +1,50 @@
+import urlcat from 'urlcat'
+import {
+    createIndicator,
+    createNextIndicator,
+    createPageable,
+    HubOptions,
+    Pageable,
+    Transaction,
+} from '@masknet/web3-shared-base'
+import { EMPTY_LIST } from '@masknet/shared-base'
+import { ChainId, isValidChainId, SchemaType } from '@masknet/web3-shared-evm'
+import type { Tx } from '../types.js'
+import { fetchFromChainbase } from './helpers.js'
+import type { HistoryAPI } from '../../entry-types.js'
+
+export class ChainbaseHistoryAPI implements HistoryAPI.Provider<ChainId, SchemaType> {
+    createTransactionFromTx(chainId: ChainId, tx: Tx): Transaction<ChainId, SchemaType> {
+        return {
+            id: tx.transaction_hash,
+            chainId,
+            status: tx.status,
+            timestamp: new Date(tx.block_timestamp).getTime(),
+            from: tx.from_address,
+            to: tx.to_address,
+            tokens: EMPTY_LIST,
+        }
+    }
+
+    async getTransactions(
+        address: string,
+        { chainId = ChainId.Mainnet, indicator }: HubOptions<ChainId> = {},
+    ): Promise<Pageable<Transaction<ChainId, SchemaType>>> {
+        if (!isValidChainId(chainId)) return createPageable(EMPTY_LIST, createIndicator(indicator))
+        const txs = await fetchFromChainbase<Tx[]>(
+            urlcat('/v1/account/txs', {
+                chainId,
+                address,
+                page: (indicator?.index ?? 0) + 1,
+            }),
+        )
+
+        if (!txs) return createPageable(EMPTY_LIST, createIndicator(indicator))
+        const assets = txs.map((x) => this.createTransactionFromTx(chainId, x))
+        return createPageable(
+            assets,
+            createIndicator(indicator),
+            assets.length ? createNextIndicator(indicator) : undefined,
+        )
+    }
+}
